leds: Add new LED driver for lm3643 chips
Currently the driver doesn't differentiate LED1 and LED2, both are
always set to the same brightness.

Also, the TX/STROBE/TORCH GPIOs are requested but not really used
since the chip allows to control everything through I2C, see
"9.5.1 Control Truth Table" in the chip documentation.

The flash has an extra timeout_ms sysfs entry to be able to modify it
easily.

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>